attempted basic intro screen, not activated yet tho
for some reason, when I fade it to black it throws an exception inside the GetCurrentFrame from Animation.h, please help cuz if we can't fix this the game won't run

+++ AlteredBeast Proto_/Project_4_Handout/Source/Animation.h
@@ -14,9 +14,8 @@ public:
 	bool pingpong = false;
 	int loopCount = 0;
 	float currentFrame = 0.0f;
+
 private:
-	
-	
 	
 	int pingpongDirection = 1;
 
@@ -31,7 +30,6 @@ public:
 	{
 		currentFrame = 0;
 	}
-
 
 
 	bool HasFinished()
@@ -56,9 +54,9 @@ public:
 	
 
 	//Changed to notconstant bc of the change in the blit function
-	 SDL_Rect& GetCurrentFrame() 
+	SDL_Rect& GetCurrentFrame() 
 	{
-		int actualFrame = currentFrame;
+		int actualFrame = (int)currentFrame;
 		if (pingpongDirection == -1)
 			actualFrame = totalFrames - currentFrame;
